Creación de catalogo de tipo de pago y avance en validaciones de reserva de areas comunes

+++ src/main/java/mx/com/admoninmuebles/service/TipoPagoServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import mx.com.admoninmuebles.dto.TipoPagoDto;
+import mx.com.admoninmuebles.error.BusinessException;
 import mx.com.admoninmuebles.persistence.model.TipoPago;
 import mx.com.admoninmuebles.persistence.repository.TipoPagoRepository;
 
@@ -42,6 +43,10 @@ public class TipoPagoServiceImpl implements TipoPagoService {
 
 	@Override
 	public void deleteById(Long idTipoPago) {
+		Optional<TipoPago> tipoPago = tipoPagoRepository.findById(idTipoPago);
+		if( !tipoPago.isPresent() ) {
+			throw new BusinessException();
+		}
 		tipoPagoRepository.deleteById(idTipoPago);
 		
 	}
